+++ rest_framework_reactive/connection.py
@@ -2,6 +2,7 @@ from __future__ import absolute_import, division, print_function, unicode_litera
 
 from django.conf import settings
 
+from channels import Group
 
 # Redis channel for receiving control messages.
 QUERYOBSERVER_REDIS_CHANNEL = 'django-rest-framework-reactive:control'
@@ -45,3 +46,21 @@ def get_queryobserver_settings():
     }
     defaults.update(getattr(settings, 'DJANGO_REST_FRAMEWORK_REACTIVE', {}))
     return defaults
+
+
+def get_subscriber_group_id(subscriber_id):
+    """Return group for specific subscriber.
+
+    :param subscriber_id: Subscriber identifier
+    :return: A channels Group
+    """
+    return 'rest_framework_reactive.subscriber.{}'.format(subscriber_id)
+
+
+def get_subscriber_group(subscriber_id):
+    """Return group for specific subscriber.
+
+    :param subscriber_id: Subscriber identifier
+    :return: A channels Group
+    """
+    return Group(get_subscriber_group_id(subscriber_id))